re-work code into components and hooks
