+++ web/src/Components/HistoryTasks.js
@@ -1,13 +1,8 @@
-import React, {useEffect, useRef, useState} from "react";
-import {fetchApplications, fetchTasks} from "../Services/Data";
-import TableCell from "@mui/material/TableCell";
+import React, {useEffect, useState} from "react";
 import Box from "@mui/material/Box";
-import KeyboardArrowUpIcon from "@mui/icons-material/KeyboardArrowUp";
-import KeyboardArrowDownIcon from "@mui/icons-material/KeyboardArrowDown";
 import Container from "@mui/material/Container";
 import Stack from "@mui/material/Stack";
 import Typography from "@mui/material/Typography";
-import Autocomplete from "@mui/material/Autocomplete";
 import TextField from "@mui/material/TextField";
 import IconButton from "@mui/material/IconButton";
 import RefreshIcon from "@mui/icons-material/Refresh";
@@ -18,15 +13,9 @@ import MenuItem from "@mui/material/MenuItem";
 import LocalizationProvider from "@mui/lab/LocalizationProvider";
 import AdapterDateFns from "@mui/lab/AdapterDateFns";
 import DatePicker from "@mui/lab/DatePicker";
-import TableContainer from "@mui/material/TableContainer";
-import Paper from "@mui/material/Paper";
-import Table from "@mui/material/Table";
-import TableHead from "@mui/material/TableHead";
-import TableRow from "@mui/material/TableRow";
-import TableBody from "@mui/material/TableBody";
-import Tooltip from "@mui/material/Tooltip";
-import {relativeTime} from "../Utils";
 import ErrorSnackbar from "./ErrorSnackbar";
+import ApplicationsFilter from "./ApplicationsFilter";
+import TasksTable, {useTasks} from "./TasksTable";
 
 const timeframes = {
   '5 minutes': 5 * 60,
@@ -38,24 +27,10 @@ const timeframes = {
   'Custom': 0,
 };
 
-const autoRefreshIntervals = {
-  '5s': 5,
-  '10s': 10,
-  '30s': 30,
-  '1m': 60,
-  'off': 0,
-};
-
 function HistoryTasks() {
-  const [tasks, setTasks] = useState([]);
-  const [applications, setApplications] = useState([]);
-
-  const [currentSort, setCurrentSort] = useState({field: "created", direction: "ASC"});
-  const [currentAutoRefresh, setCurrentAutoRefresh] = useState(autoRefreshIntervals['30s']);
-  const autoRefreshIntervalRef = useRef(null);
-  const [currentApplication, setCurrentApplication] = useState(null);
-
   const [loadingError, setLoadingError] = useState(null);
+  const {tasks, sortField, setSortField, refreshTasks} = useTasks({ setLoadingError });
+  const [currentApplication, setCurrentApplication] = useState(null);
   const [currentTimeframe, setCurrentTimeframe] = useState(timeframes['1 hour']);
   const [currentCustomTimestamp, setCurrentCustomTimestamp] = useState(null);
 
@@ -69,55 +44,6 @@ function HistoryTasks() {
     }
   }
 
-  const refreshTasks = (timeframe, application) => {
-    let timestamp = Math.floor(Date.now() / 1000) - timeframe;
-    // get tasks by timestamp
-    fetchTasks(timestamp, application)
-        .then(items => { setTasksSorted(items, currentSort); })
-        .catch(error => { setLoadingError(error.message); });
-  };
-
-  const setTasksSorted = (tasks, sort) => {
-    // sort tasks
-    tasks.sort((a, b) => {
-      let aField = a[sort.field];
-      let bField = b[sort.field];
-      if (aField === bField) {
-        return 0;
-      }
-      if (aField > bField) {
-        return sort.direction === 'ASC' ? -1 : 1;
-      } else {
-        return sort.direction === 'ASC' ? 1 : -1;
-      }
-    });
-
-    // save sorted tasks
-    setTasks([].concat(tasks));
-  };
-
-  const triggerSortChange = (field) => {
-    // change sort parameters
-    let sortFieldChange = {...currentSort};
-    if (sortFieldChange.field === field) {
-      sortFieldChange.direction = sortFieldChange.direction === 'ASC' ? 'DESC' : 'ASC';
-    } else {
-      sortFieldChange.field = field;
-      sortFieldChange.direction = 'ASC';
-    }
-    setCurrentSort(sortFieldChange);
-    // set sorted tasks
-    setTasksSorted(tasks, sortFieldChange);
-  };
-
-  // initial load
-  useEffect(() => {
-    refreshTasks(currentTimeframe, currentApplication);
-    fetchApplications()
-        .then(items => { setApplications(items) })
-        .catch(error => { setLoadingError(error.message); });
-  }, []);
-
   const handleTimeframeChange = (event) => {
     let newTimeframe = event.target.value;
     setCurrentTimeframe(newTimeframe);
@@ -126,11 +52,6 @@ function HistoryTasks() {
     } else if (currentCustomTimestamp !== null) {
       refreshTasks(Math.floor((Date.now() - currentCustomTimestamp.getTime()) / 1000), currentApplication);
     }
-  };
-
-  const handleApplicationsChange = (event, newValue) => {
-    setCurrentApplication(newValue);
-    refreshTasks(calculateCurrentTimeframe(), newValue);
   };
 
   const handleCustomTimestampChange = (newValue) => {
@@ -146,16 +67,9 @@ function HistoryTasks() {
     }
   };
 
-  const TableCellSorted = ({field, children}) => {
-    return <TableCell
-        onClick={() => {  triggerSortChange(field); }}
-        sx={{cursor: "pointer",}}
-    >
-      <Box sx={{display: "flex", flexDirection: "row",}}>
-        {children} {currentSort.field === field && (currentSort.direction === "ASC" ? <KeyboardArrowUpIcon /> : <KeyboardArrowDownIcon />)}
-      </Box>
-    </TableCell>
-  };
+  useEffect(() => {
+    refreshTasks(currentTimeframe, currentApplication);
+  }, []);
 
   return (
     <Container maxWidth="xl">
@@ -164,14 +78,13 @@ function HistoryTasks() {
           History tasks
         </Typography>
         <Box>
-          <Autocomplete
-              size={"small"}
-              disablePortal
-              options={applications}
-              sx={{ width: 220 }}
-              renderInput={(params) => <TextField {...params} label="Application" />}
-              value={currentApplication || null}
-              onChange={handleApplicationsChange}
+          <ApplicationsFilter
+              value={currentApplication}
+              onChange={(value) => {
+                setCurrentApplication(value);
+                refreshTasks(calculateCurrentTimeframe(), value);
+              }}
+              setLoadingError={setLoadingError}
           />
         </Box>
         {currentTimeframe === timeframes['Custom'] &&  <Box sx={{width: 200}}>
@@ -206,58 +119,11 @@ function HistoryTasks() {
           <RefreshIcon/>
         </IconButton>
       </Stack>
-      <TableContainer component={Paper}>
-        <Table sx={{minWidth: 650}} aria-label="simple table">
-          <TableHead>
-            <TableRow>
-              <TableCellSorted field={"id"}>ID</TableCellSorted>
-              <TableCellSorted field={"app"}>Application</TableCellSorted>
-              <TableCellSorted field={"project"}>Project</TableCellSorted>
-              <TableCellSorted field={"author"}>Author</TableCellSorted>
-              <TableCellSorted field={"status"}>Status</TableCellSorted>
-              <TableCellSorted field={"created"}>Started</TableCellSorted>
-              <TableCellSorted field={"updated"}>Updated</TableCellSorted>
-              <TableCellSorted field={"images"}>Images</TableCellSorted>
-            </TableRow>
-          </TableHead>
-          <TableBody>
-            {tasks.map((task) => (
-                <TableRow
-                    key={task.id}
-                    sx={{'&:last-child td, &:last-child th': {border: 0}}}
-                >
-                  <TableCell component="th" scope="row">
-                    {task.id}
-                  </TableCell>
-                  <TableCell>{task.app}</TableCell>
-                  <TableCell>{task.project}</TableCell>
-                  <TableCell>{task.author}</TableCell>
-                  <TableCell>{task.status}</TableCell>
-                  <TableCell>
-                    <Tooltip title={new Date(task.created * 1000).toISOString()}>
-                      <span>{relativeTime(task.created * 1000)}</span>
-                    </Tooltip>
-                  </TableCell>
-                  <TableCell>
-                    <Tooltip title={new Date(task.updated * 1000).toISOString()}>
-                      <span>{relativeTime(task.updated * 1000)}</span>
-                    </Tooltip>
-                  </TableCell>
-                  <TableCell>
-                    {task.images.map((item, index) => {
-                      return <div key={index}>{item.image}:{item.tag}</div>
-                    })}
-                  </TableCell>
-                </TableRow>
-            ))}
-            {tasks.length === 0 && <TableRow>
-              <TableCell colSpan={100} sx={{textAlign: "center"}}>
-                No tasks were found within provided timeframe
-              </TableCell>
-            </TableRow>}
-          </TableBody>
-        </Table>
-      </TableContainer>
+      <TasksTable
+          tasks={tasks}
+          sortField={sortField}
+          setSortField={setSortField}
+      />
       <ErrorSnackbar message={loadingError} setMessage={setLoadingError}/>
     </Container>
   );
