+++ web/src/Components/RecentTasks.js
@@ -1,32 +1,17 @@
 import React, {useEffect, useRef, useState} from "react";
-import {fetchApplications, fetchTasks} from "../Services/Data";
-import TableCell from "@mui/material/TableCell";
 import Box from "@mui/material/Box";
-import KeyboardArrowUpIcon from "@mui/icons-material/KeyboardArrowUp";
-import KeyboardArrowDownIcon from "@mui/icons-material/KeyboardArrowDown";
 import Container from "@mui/material/Container";
 import Stack from "@mui/material/Stack";
 import Typography from "@mui/material/Typography";
-import Autocomplete from "@mui/material/Autocomplete";
-import TextField from "@mui/material/TextField";
 import IconButton from "@mui/material/IconButton";
 import RefreshIcon from "@mui/icons-material/Refresh";
 import FormControl from "@mui/material/FormControl";
 import InputLabel from "@mui/material/InputLabel";
 import Select from "@mui/material/Select";
 import MenuItem from "@mui/material/MenuItem";
-import LocalizationProvider from "@mui/lab/LocalizationProvider";
-import AdapterDateFns from "@mui/lab/AdapterDateFns";
-import DatePicker from "@mui/lab/DatePicker";
-import TableContainer from "@mui/material/TableContainer";
-import Paper from "@mui/material/Paper";
-import Table from "@mui/material/Table";
-import TableHead from "@mui/material/TableHead";
-import TableRow from "@mui/material/TableRow";
-import TableBody from "@mui/material/TableBody";
-import Tooltip from "@mui/material/Tooltip";
-import {relativeTime} from "../Utils";
 import ErrorSnackbar from "./ErrorSnackbar";
+import ApplicationsFilter from "./ApplicationsFilter";
+import TasksTable, {useTasks} from "./TasksTable";
 
 const autoRefreshIntervals = {
   '5s': 5,
@@ -37,64 +22,16 @@ const autoRefreshIntervals = {
 };
 
 function RecentTasks() {
-  const [tasks, setTasks] = useState([]);
-  const [applications, setApplications] = useState([]);
-
-  const [currentSort, setCurrentSort] = useState({field: "created", direction: "ASC"});
+  const [loadingError, setLoadingError] = useState(null);
+  const {tasks, sortField, setSortField, refreshTasks} = useTasks({ setLoadingError });
   const [currentAutoRefresh, setCurrentAutoRefresh] = useState(autoRefreshIntervals['30s']);
   const autoRefreshIntervalRef = useRef(null);
   const [currentApplication, setCurrentApplication] = useState(null);
-
-  const [loadingError, setLoadingError] = useState(null);
   const currentTimeframe = 60 * 60;
-
-  const refreshTasks = (application) => {
-    let timestamp = Math.floor(Date.now() / 1000) - currentTimeframe;
-    // get tasks by timestamp
-    fetchTasks(timestamp, application)
-        .then(items => { setTasksSorted(items, currentSort); })
-        .catch(error => { setLoadingError(error.message); });
-  };
-
-  const setTasksSorted = (tasks, sort) => {
-    // sort tasks
-    tasks.sort((a, b) => {
-      let aField = a[sort.field];
-      let bField = b[sort.field];
-      if (aField === bField) {
-        return 0;
-      }
-      if (aField > bField) {
-        return sort.direction === 'ASC' ? -1 : 1;
-      } else {
-        return sort.direction === 'ASC' ? 1 : -1;
-      }
-    });
-
-    // save sorted tasks
-    setTasks([].concat(tasks));
-  };
-
-  const triggerSortChange = (field) => {
-    // change sort parameters
-    let sortFieldChange = {...currentSort};
-    if (sortFieldChange.field === field) {
-      sortFieldChange.direction = sortFieldChange.direction === 'ASC' ? 'DESC' : 'ASC';
-    } else {
-      sortFieldChange.field = field;
-      sortFieldChange.direction = 'ASC';
-    }
-    setCurrentSort(sortFieldChange);
-    // set sorted tasks
-    setTasksSorted(tasks, sortFieldChange);
-  };
 
   // initial load
   useEffect(() => {
-    refreshTasks(currentApplication);
-    fetchApplications()
-        .then(items => { setApplications(items) })
-        .catch(error => { setLoadingError(error.message); });
+    refreshTasks(currentTimeframe, currentApplication);
   }, []);
 
   // we reset interval on any state change (because we use the state variables for data retrieval)
@@ -108,7 +45,7 @@ function RecentTasks() {
     }
     // set interval
     autoRefreshIntervalRef.current = setInterval(() => {
-      refreshTasks(currentApplication);
+      refreshTasks(currentTimeframe, currentApplication);
     }, currentAutoRefresh * 1000);
 
     // clear interval on exit
@@ -123,22 +60,6 @@ function RecentTasks() {
     setCurrentAutoRefresh(event.target.value);
   };
 
-  const handleApplicationsChange = (event, newValue) => {
-    setCurrentApplication(newValue);
-    refreshTasks(newValue);
-  };
-
-  const TableCellSorted = ({field, children}) => {
-    return <TableCell
-        onClick={() => {  triggerSortChange(field); }}
-        sx={{cursor: "pointer",}}
-    >
-      <Box sx={{display: "flex", flexDirection: "row",}}>
-        {children} {currentSort.field === field && (currentSort.direction === "ASC" ? <KeyboardArrowUpIcon /> : <KeyboardArrowDownIcon />)}
-      </Box>
-    </TableCell>
-  };
-
   return (
     <Container maxWidth="xl">
       <Stack direction="row" spacing={2} alignItems="center">
@@ -146,14 +67,13 @@ function RecentTasks() {
           Recent tasks
         </Typography>
         <Box>
-          <Autocomplete
-              size={"small"}
-              disablePortal
-              options={applications}
-              sx={{ width: 220 }}
-              renderInput={(params) => <TextField {...params} label="Application" />}
-              value={currentApplication || null}
-              onChange={handleApplicationsChange}
+          <ApplicationsFilter
+            value={currentApplication}
+            onChange={(value) => {
+              setCurrentApplication(value);
+              refreshTasks(currentTimeframe, value);
+            }}
+            setLoadingError={setLoadingError}
           />
         </Box>
         <Box sx={{minWidth: 120}}>
@@ -177,58 +97,11 @@ function RecentTasks() {
           <RefreshIcon/>
         </IconButton>
       </Stack>
-      <TableContainer component={Paper}>
-        <Table sx={{minWidth: 650}} aria-label="simple table">
-          <TableHead>
-            <TableRow>
-              <TableCellSorted field={"id"}>ID</TableCellSorted>
-              <TableCellSorted field={"app"}>Application</TableCellSorted>
-              <TableCellSorted field={"project"}>Project</TableCellSorted>
-              <TableCellSorted field={"author"}>Author</TableCellSorted>
-              <TableCellSorted field={"status"}>Status</TableCellSorted>
-              <TableCellSorted field={"created"}>Started</TableCellSorted>
-              <TableCellSorted field={"updated"}>Updated</TableCellSorted>
-              <TableCellSorted field={"images"}>Images</TableCellSorted>
-            </TableRow>
-          </TableHead>
-          <TableBody>
-            {tasks.map((task) => (
-                <TableRow
-                    key={task.id}
-                    sx={{'&:last-child td, &:last-child th': {border: 0}}}
-                >
-                  <TableCell component="th" scope="row">
-                    {task.id}
-                  </TableCell>
-                  <TableCell>{task.app}</TableCell>
-                  <TableCell>{task.project}</TableCell>
-                  <TableCell>{task.author}</TableCell>
-                  <TableCell>{task.status}</TableCell>
-                  <TableCell>
-                    <Tooltip title={new Date(task.created * 1000).toISOString()}>
-                      <span>{relativeTime(task.created * 1000)}</span>
-                    </Tooltip>
-                  </TableCell>
-                  <TableCell>
-                    <Tooltip title={new Date(task.updated * 1000).toISOString()}>
-                      <span>{relativeTime(task.updated * 1000)}</span>
-                    </Tooltip>
-                  </TableCell>
-                  <TableCell>
-                    {task.images.map((item, index) => {
-                      return <div key={index}>{item.image}:{item.tag}</div>
-                    })}
-                  </TableCell>
-                </TableRow>
-            ))}
-            {tasks.length === 0 && <TableRow>
-              <TableCell colSpan={100} sx={{textAlign: "center"}}>
-                No tasks were found within provided timeframe
-              </TableCell>
-            </TableRow>}
-          </TableBody>
-        </Table>
-      </TableContainer>
+      <TasksTable
+          tasks={tasks}
+          sortField={sortField}
+          setSortField={setSortField}
+      />
       <ErrorSnackbar message={loadingError} setMessage={setLoadingError}/>
     </Container>
   );
